Add files via upload
I have finally reorganised the project. Originally, I was too dependent on visual studio's "Solution View" in which the files are linked. However, even if those files are linked, they were not organised as per the solution view and I had to sort them into the appropriate folders (Header or Source files). I had also separated the Classes that had their functions in the header file into both the header file that contains declarations and the source files for the functions.